+++ src/lang/nodes/Placeholder.js
@@ -2,7 +2,7 @@ export function suggest(node, ctx) {
     ctx.queryRoot(node.range[1]);
 }
 export function compile(node, ctx) {
-    ctx.put('$');
+    ctx.put(ctx.scope.$ref);
 }
 export function walk() {}
 export function stringify() {}